requirements.txt
streamlit==1.39.0
langchain-google-genai==2.0.5
langchain-core==0.3.28
python-dotenv==1.0.1
